Move VFP tables to Schedule.
